pyproject.toml
[tool.poetry]
name = "terramar"
version = "0.1.0"
description = ""
authors = ["adria0 <nowhere@>"]

[tool.poetry.dependencies]
python = "^3.9"
pygame = "^2.1.2"
"RPi.GPIO" = "^0.7.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
